kafka-docker/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Flask==3.1.2
idna==3.11
implicit==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kafka-python==2.2.15
MarkupSafe==3.0.3
mypy_extensions==1.1.0
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pandera==0.26.1
psutil==7.1.0
pyarrow==21.0.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scikit-surprise==1.1.4
scipy==1.16.2
six==1.17.0
surprise==0.1
threadpoolctl==3.6.0
tqdm==4.67.1
typeguard==4.4.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
